requirements.txt
Django>=5.2.1
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-ninja>=1.4.1
django-filter>=23.5
coreapi>=2.3.3
pydantic>=2.5
psycopg2-binary>=2.9.10
uvicorn>=0.34.2
